requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==4.1.1
django-debug-toolbar==3.7.0
django-heroku==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.1
gunicorn==20.1.0
psycopg2==2.9.3
PyJWT==2.5.0
pytz==2022.4
sqlparse==0.4.3
tzdata==2022.4
whitenoise==6.2.0
